completada interfaz para gestión de grupos de sufijos. Agregada confirmación para eliminar items en todas las listas

+++ tweetfilter/static/js/filtering/blocked_user_settings.js
@@ -77,7 +77,9 @@ var
                     );
 
                     $('#delete_blocked_user_' + elem.id).click(function () {
-                        delete_blocked_user(elem.id);
+                        if (confirm("Está seguro de que desea desbloquear al usuario?")) {
+                            delete_blocked_user(elem.id);
+                        }
                     });
                 });
